dom/requestsync/tests/test_bug1151082.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for RequestSync bug 1151082</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="common_basic.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script type="application/javascript;version=1.7">


  var tests = [
    function() {
      SpecialPowers.pushPrefEnv({"set": [["dom.requestSync.enabled", true],
                                         ["dom.requestSync.minInterval", 1],
                                         ["dom.ignore_webidl_scope_checks", true]]}, runTests);
    },

    function() {
      SpecialPowers.pushPermissions(
        [{ "type": "requestsync-manager", "allow": 1, "context": document } ], runTests);
    },

    function() {
      SpecialPowers.importInMainProcess("resource://gre/modules/RequestSyncService.jsm");
      runTests();
    },

    function() {
      counter = 2;
      function registerCb() {
        if (!--counter) {
          ok(true, "All the registrations are done.");
          runTests();
        }
      }

      navigator.sync.register('foobar', { minInterval: 5, wakeUpPage:'/' }).then(registerCb, genericError);
      navigator.sync.register('barfoo', { minInterval: 5, wakeUpPage:'/' }).then(registerCb, genericError);
    },

    function() {
      counter = 2;
      function unregisterCb() {
        if (!--counter) {
          ok(true, "All the unregistrations are done.");
          runTests();
        }
      }

      navigator.sync.unregister('foobar').then(unregisterCb, genericError);
      navigator.sync.unregister('barfoo').then(unregisterCb, genericError);
    }
  ];

  function runTests() {
    if (!tests.length) {
      finish();
      return;
    }

    var test = tests.shift();
    test();
  }

  function finish() {
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  runTests();

  </script>
</body>
</html>
